Moved form and added in sidebar
