+++ services/frontend-v3/src/components/profileForms/employmentDataForm/EmploymentDataFormView.jsx
@@ -54,7 +54,6 @@ const EmploymentDataFormView = ({
   intl,
   history,
   userId,
-  employmentEquityOptions,
 }) => {
   const axios = useAxios();
   const [form] = Form.useForm();
@@ -257,7 +256,6 @@ const EmploymentDataFormView = ({
         actingEndDate: profile.actingStartDate
           ? moment(profile.actingStartDate)
           : undefined,
-        employmentEquityGroups: profile.employmentEquityGroups,
       };
     }
     return {};
@@ -671,39 +669,7 @@ const EmploymentDataFormView = ({
               {getTempRoleForm(displayActingRoleForm)}
             </Col>
           </Row>
-          <Divider style={styles.headerDiv} />
-          <Row
-            justify="space-between"
-            style={styles.sectionHeader}
-            align="middle"
-          >
-            <Title level={3} style={styles.formTitle}>
-              <FormattedMessage id="profile.employment.equity.groups" />
-            </Title>
-            <CardVisibilityToggle
-              visibleCards={profileInfo.visibleCards}
-              cardName="employmentEquityGroup"
-              type="form"
-            />
-          </Row>
-          <Row gutter={24}>
-            <Col className="gutter-row" span={24}>
-              <Form.Item name="employmentEquityGroups">
-                <Select
-                  showSearch
-                  mode="multiple"
-                  placeholder={<FormattedMessage id="setup.select" />}
-                  allowClear
-                  filterOption={filterOption}
-                  className="custom-bubble-select-style"
-                >
-                  {employmentEquityOptions.map(({ key, text }) => (
-                    <Option key={key}>{text}</Option>
-                  ))}
-                </Select>
-              </Form.Item>
-            </Col>
-          </Row>
+          
           <Divider style={styles.headerDiv} />
           <Row
             justify="space-between"
@@ -759,7 +725,6 @@ EmploymentDataFormView.propTypes = {
   intl: IntlPropType,
   history: HistoryPropType.isRequired,
   userId: PropTypes.string.isRequired,
-  employmentEquityOptions: KeyTitleOptionsPropType.isRequired,
 };
 
 EmploymentDataFormView.defaultProps = {
