+++ services/frontend-v3/src/components/profileForms/primaryInfoForm/PrimaryInfoForm.jsx
@@ -2,6 +2,7 @@ import React, { useState, useEffect, useCallback } from "react";
 import PropTypes from "prop-types";
 import { useHistory } from "react-router-dom";
 import { useSelector } from "react-redux";
+import { useIntl } from "react-intl";
 import useAxios from "../../../utils/axios-instance";
 import PrimaryInfoFormView from "./PrimaryInfoFormView";
 import handleError from "../../../functions/handleError";
@@ -10,12 +11,14 @@ const PrimaryInfoForm = ({ formType }) => {
   const [locationOptions, setLocationOptions] = useState([]);
   const [profileInfo, setProfileInfo] = useState(null);
   const [load, setLoad] = useState(false);
+  const [employmentEquityOptions, setEmploymentEquityOptions] = useState([]);
 
   const { id, email } = useSelector((state) => state.user);
   const { locale } = useSelector((state) => state.settings);
   const axios = useAxios();
 
   const history = useHistory();
+  const intl = useIntl();
 
   // Get possible locations for form drop down
   const getLocations = useCallback(async () => {
@@ -53,6 +56,31 @@ const PrimaryInfoForm = ({ formType }) => {
       .then(() => setLoad(true));
   }, [getLocations, getProfileInfo]);
 
+  useEffect(() => {
+    setEmploymentEquityOptions([
+      {
+        key: "WOMEN",
+        value: "WOMEN",
+        text: intl.formatMessage({ id: "employment.equity.group.woman" }),
+      },
+      {
+        key: "INDIGENOUS",
+        value: "INDIGENOUS",
+        text: intl.formatMessage({ id: "employment.equity.group.indigenous" }),
+      },
+      {
+        key: "DISABILITY",
+        value: "DISABILITY",
+        text: intl.formatMessage({ id: "employment.equity.group.disability" }),
+      },
+      {
+        key: "MINORITY",
+        value: "MINORITY",
+        text: intl.formatMessage({ id: "employment.equity.group.minority" }),
+      },
+    ]);
+  }, [intl]);
+
   return (
     <PrimaryInfoFormView
       locationOptions={locationOptions}
@@ -62,6 +90,7 @@ const PrimaryInfoForm = ({ formType }) => {
       history={history}
       userId={id}
       email={email}
+      employmentEquityOptions={employmentEquityOptions}
     />
   );
 };
